no log: Typescript 7 and oxlint
