+++ frontend/.oxlintrc.json
@@ -1,0 +1,110 @@
+{
+  "$schema": "./node_modules/oxlint/configuration_schema.json",
+  "plugins": ["react-hooks"],
+  "categories": {
+    "correctness": "error"
+  },
+  "jsPlugins": [
+    "eslint-plugin-simple-import-sort",
+    "eslint-plugin-testing-library",
+    "eslint-plugin-react-refresh",
+    {
+      "name": "bazarr-custom",
+      "specifier": "./oxlint-plugins/bazarr-custom.ts"
+    }
+  ],
+  "ignorePatterns": [
+    "dist",
+    "build",
+    "coverage",
+    "node_modules",
+    "dev-dist",
+    ".husky",
+    "vite.config.ts"
+  ],
+  "rules": {
+    "react-hooks/rules-of-hooks": "error",
+    "react-hooks/exhaustive-deps": "error",
+    "react-refresh/only-export-components": [
+      "error",
+      { "allowConstantExport": true, "extraHOCs": ["withModal"] }
+    ],
+    "no-console": "error",
+    "bazarr-custom/camelcase": "error",
+    "no-restricted-imports": ["error", { "patterns": ["../*"] }],
+    "no-empty-function": "error",
+    "no-unused-expressions": "error",
+    "no-unused-vars": "error",
+    "no-useless-rename": "error",
+    "bazarr-custom/no-let": "error",
+    "bazarr-custom/no-function-declaration": "error",
+    "simple-import-sort/imports": [
+      "error",
+      {
+        "groups": [
+          [
+            "^react",
+            "^@mantine/",
+            "^(\\w|@\\w)",
+            "^\\u0000",
+            "^@/\\w",
+            "^\\.\\.(?!/?$)",
+            "^\\.\\./?$",
+            "^\\./(?=.*/)(?!/?$)",
+            "^\\.(?!/?$)",
+            "^\\./?$",
+            "^.+\\.?(css)$"
+          ]
+        ]
+      }
+    ]
+  },
+  "overrides": [
+    {
+      "files": [
+        "src/components/index.tsx",
+        "src/pages/Settings/components/index.tsx",
+        "src/tests/index.tsx"
+      ],
+      "rules": {
+        "react-refresh/only-export-components": "off"
+      }
+    },
+    {
+      "files": [
+        "**/__tests__/**/*.{js,jsx,ts,tsx}",
+        "**/*.{test,spec}.{js,jsx,ts,tsx}"
+      ],
+      "rules": {
+        "testing-library/await-async-events": [
+          "error",
+          { "eventModule": "userEvent" }
+        ],
+        "testing-library/await-async-queries": "error",
+        "testing-library/await-async-utils": "error",
+        "testing-library/no-await-sync-events": [
+          "error",
+          { "eventModules": ["fire-event"] }
+        ],
+        "testing-library/no-await-sync-queries": "error",
+        "testing-library/no-container": "error",
+        "testing-library/no-debugging-utils": "warn",
+        "testing-library/no-dom-import": ["error", "react"],
+        "testing-library/no-global-regexp-flag-in-query": "error",
+        "testing-library/no-manual-cleanup": "error",
+        "testing-library/no-node-access": "error",
+        "testing-library/no-promise-in-fire-event": "error",
+        "testing-library/no-render-in-lifecycle": "error",
+        "testing-library/no-unnecessary-act": "error",
+        "testing-library/no-wait-for-multiple-assertions": "error",
+        "testing-library/no-wait-for-side-effects": "error",
+        "testing-library/no-wait-for-snapshot": "error",
+        "testing-library/prefer-find-by": "error",
+        "testing-library/prefer-presence-queries": "error",
+        "testing-library/prefer-query-by-disappearance": "error",
+        "testing-library/prefer-screen-queries": "error",
+        "testing-library/render-result-naming-convention": "error"
+      }
+    }
+  ]
+}
