+++ frontend/package.json
@@ -42,7 +42,6 @@
     "workbox-build": "^7.4.1"
   },
   "devDependencies": {
-    "@eslint/js": "^10.0.1",
     "@fontsource/roboto": "^5.2.10",
     "@fortawesome/fontawesome-svg-core": "^7.3.0",
     "@fortawesome/free-brands-svg-icons": "^7.3.0",
@@ -59,23 +58,19 @@
     "@types/node": "^26.1.1",
     "@types/react": "^19.2.17",
     "@types/react-dom": "^19.2.3",
-    "@typescript-eslint/eslint-plugin": "^8.64.0",
-    "@typescript-eslint/parser": "^8.64.0",
     "@vite-pwa/assets-generator": "^1.0.2",
     "@vitejs/plugin-react": "^6.0.5",
     "@vitest/coverage-v8": "^4.1.10",
     "@vitest/ui": "^4.1.10",
     "clsx": "^2.1.1",
-    "eslint": "^10.7.0",
-    "eslint-plugin-react-hooks": "^7.1.1",
     "eslint-plugin-react-refresh": "^0.5.3",
     "eslint-plugin-simple-import-sort": "^14.0.0",
     "eslint-plugin-testing-library": "^7.16.2",
-    "globals": "^17.7.0",
     "husky": "^9.1.7",
     "jsdom": "^30.0.1",
     "lodash": "^4.18.1",
     "msw": "^2.15.0",
+    "oxlint": "^1.78.0",
     "postcss-preset-mantine": "^1.18.0",
     "postcss-simple-vars": "^7.0.1",
     "prettier": "^3.9.5",
@@ -86,8 +81,7 @@
     "stylelint": "^17.14.1",
     "stylelint-config-css-modules": "^4.6.0",
     "stylelint-config-standard-scss": "^17.0.0",
-    "typescript": "^5.9.3",
-    "typescript-eslint": "^8.64.0",
+    "typescript": "^7.0.2",
     "vite": "^8.1.4",
     "vite-plugin-checker": "^0.14.4",
     "vite-plugin-pwa": "^1.3.0",
@@ -97,8 +91,8 @@
   "scripts": {
     "build": "vite build",
     "build:ci": "vite build -m development",
-    "check": "eslint src",
-    "check:fix": "eslint src --fix",
+    "check": "oxlint src",
+    "check:fix": "oxlint src --fix",
     "check:ts": "tsc --noEmit --incremental false",
     "check:fmt": "prettier -c .",
     "check:style": "stylelint \"src/**/*.{css,scss}\" --max-warnings 0",
