+++ frontend/src/App/Navbar.tsx
@@ -33,8 +33,8 @@ import { useSettingsMutation, useSystemSettings } from "@/apis/hooks";
 import { QueryKeys } from "@/apis/queries/keys";
 import { Action } from "@/components";
 import { useNavbar } from "@/contexts/Navbar";
-import { useRouteItems } from "@/Router";
 import { CustomRouteObject, Route } from "@/Router/type";
+import { useRouteItems } from "@/Router/useRouteItems";
 import { BuildKey, pathJoin } from "@/utilities";
 import { LOG } from "@/utilities/console";
 import styles from "./Navbar.module.scss";
