+++ frontend/src/Router/useRouteItems.ts
@@ -1,0 +1,6 @@
+import { createContext, useContext } from "react";
+import { CustomRouteObject } from "@/Router/type";
+
+export const RouterItemContext = createContext<CustomRouteObject[]>([]);
+
+export const useRouteItems = () => useContext(RouterItemContext);
