+++ frontend/src/components/spotlight.ts
@@ -1,0 +1,3 @@
+import { createSpotlight } from "@mantine/spotlight";
+
+export const [spotlightStore, spotlightApi] = createSpotlight();
