+++ frontend/src/pages/Series/Editor.tsx
@@ -53,7 +53,7 @@ const SeriesMassEditor: FunctionComponent = () => {
         accessorKey: "audioLanguage",
         cell: ({
           row: {
-            original: { audioLanguage: audioLanguage },
+            original: { audioLanguage },
           },
         }) => {
           return <AudioList audios={audioLanguage}></AudioList>;
