+++ frontend/src/pages/Settings/utilities/useSettings.ts
@@ -1,0 +1,9 @@
+import { createContext, useContext } from "react";
+
+export const SettingsContext = createContext<Settings | null>(null);
+
+export const useSettings = () => {
+  const context = useContext(SettingsContext);
+
+  return context;
+};
