+++ frontend/tsconfig.json
@@ -15,7 +15,6 @@
     "isolatedModules": true,
     "noEmit": true,
     "jsx": "preserve",
-    "baseUrl": "./",
     "incremental": true,
     "noFallthroughCasesInSwitch": true,
     "paths": {
